requirements.txt
openpyxl==3.1.5
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.4
brotli==1.2.0
celery==5.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cssselect2==0.9.0
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.6
django-allauth==65.12.1
django-celery-beat==2.8.0
django-cors-headers==4.6.0
django-filter==23.5
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.61.1
google-auth==2.47.0
google-auth-oauthlib==1.2.4
gunicorn==23.0.0
html5lib==1.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
oauthlib==3.3.1
packaging==25.0
pillow==11.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydyf==0.12.1
PyJWT==2.10.1
pyphen==0.17.2
pytest==7.4.4
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0
python-dotenv==1.2.1
razorpay==2.0.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.5.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.5.0
vine==5.1.0
wcwidth==0.6.0
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.11.0
zopfli==0.4.1
